Separate tests into deterministic and hypothesis-based
