+++ tests/test_deterministic/test_constants.py
