+++ tests/test_hypothesis/__init__.py
@@ -1,0 +1,9 @@
+"""
+Initializations for the tests.
+"""
+# isort: THIRDPARTY
+from hypothesis import HealthCheck, settings
+
+settings.register_profile(
+    "tracing", deadline=None, suppress_health_check=[HealthCheck.too_slow]
+)
